move local_span to range.hpp
