+++ include/ginkgo/core/base/block_operator.hpp
@@ -9,6 +9,7 @@
 #include <ginkgo/core/base/dense_cache.hpp>
 #include <ginkgo/core/base/lin_op.hpp>
 #include <ginkgo/core/base/polymorphic_object.hpp>
+#include <ginkgo/core/base/range.hpp>
 
 
 namespace gko {
